Add Rust FFI library to the repo
For the moment, the libraries are committed into the correct place in
the repo to facilitate development of the internal interface to the Rust
library. More work will be required later to automate updating this
library with new releases, and the library should probably be removed
from the repo at that point.